code optimisation

+++ controller/product.controller.js
@@ -1,6 +1,5 @@
 const Product = require('../model/product');
 const productValidator = require('../validator/product.validator');
-const mongoose = require('mongoose');
 
 exports.add = async (req, res) => {
   try {
@@ -20,7 +19,7 @@ exports.add = async (req, res) => {
           stock: stock,
           status: status
         })
-        res.status(200).json({'_id': product._id});
+        res.status(200).json({product: product});
     }
   } catch (err) {
     res.status(400).send(err);
@@ -30,12 +29,15 @@ exports.add = async (req, res) => {
 exports.update = async (req, res) => {
   try {
     const { title, type, description, price, rating, stock, status } = req.body;
-    const { _id } = req.query
-    const validateResult = productValidator.validate(req.body);
+    const _id = req.body._id || req.query._id || req.headers["_id"];
+    if (_id === undefined) {
+      throw {message: "Product id required!"};
+    }
+    const validateResult = productValidator.validate({ title, type, description, price, rating, stock, status });
     if (validateResult.error) {
       throw validateResult.error.message;
     } else {
-      var query = { '_id': mongoose.Types.ObjectId(_id) };
+      var query = { '_id': _id };
       var newData = { title: title, type: type, description: description, price: price, rating: rating, stock: stock, status: status };
       
       await Product.findOneAndUpdate(query, { $set: newData }, { new: true });
@@ -50,7 +52,10 @@ exports.update = async (req, res) => {
 
 exports.remove = async (req, res) => {
   try {
-    const { _id } = req.query;
+    const _id = req.body._id || req.query._id || req.headers["_id"];
+    if (_id === undefined) {
+      throw {message: "Product id required!"};
+    }
     await Product.deleteOne({ _id: _id });
     res.status(200).json({'message': 'success'});
   } catch (error) {
@@ -70,10 +75,10 @@ exports.getAll = async (req, res) => {
 exports.uploadProductPic = async (req, res) => {
   try {
     const { name, _id, url } = req;
-    var query = { '_id': mongoose.Types.ObjectId(_id) };
+    var query = { '_id': _id };
     var newData = { pic: name };
     await Product.findOneAndUpdate(query, { $set: newData }, { new: true });
-    res.status(200).send({ "message": 'success', file: 'name', url: url });
+    res.status(200).send({ "message": 'success', file: name, url: url });
   } catch (error) {
     res.status(400).send(error);
   }
@@ -82,7 +87,7 @@ exports.uploadProductPic = async (req, res) => {
 exports.uploadProductPics = async (req, res) => {
   try {
     const { _id, images } = req;
-    var query = { '_id': mongoose.Types.ObjectId(_id) };
+    var query = { '_id': _id };
     var newData = { images: images };
     const productUpdate = await Product.findOneAndUpdate(query, { $set: newData }, { new: true });
     res.status(200).send({ "message": 'success', 'images': images });
